Refactor world_generation.rs to modules
