+++ src/item.rs
@@ -3,11 +3,10 @@ use std::fmt;
 use crate::animations::{AnimationPosTracker, AttackAnimationTimer};
 use crate::assets::Graphics;
 use crate::attributes::{AttributeChangeEvent, BlockAttributeBundle, Health, ItemAttributes};
+use crate::combat::ObjBreakEvent;
 use crate::inventory::{Inventory, InventoryItemStack, ItemStack};
 use crate::ui::InventoryState;
-use crate::world_generation::{
-    ChunkObjectData, TileMapPositionData, WorldObjectEntityData, CHUNK_SIZE,
-};
+use crate::world::{ChunkObjectData, TileMapPositionData, WorldObjectEntityData, CHUNK_SIZE};
 use crate::{AnimationTimer, GameParam, GameState, Limb, YSort};
 use bevy::prelude::*;
 use bevy::sprite::MaterialMesh2dBundle;
@@ -411,6 +410,7 @@ impl WorldObject {
                     durability: durability_tooltip,
                 })
                 .insert(Equipment(limb))
+                .insert(YSort)
                 .insert(Name::new("EquipItem"))
                 .insert(self)
                 .id();
@@ -488,6 +488,7 @@ impl WorldObject {
             .insert(Equipment(limb))
             .insert(MainHand)
             .insert(Name::new("HeldItem"))
+            .insert(YSort)
             .insert(inv_item_stack.item_stack.attributes.clone())
             .insert(self)
             .set_parent(player_e)
@@ -610,104 +611,44 @@ impl WorldObject {
             .insert(item, (stack.clone(), transform));
         item
     }
-    pub fn attempt_to_break_item(
-        self,
-        commands: &mut Commands,
-        game: &mut GameParam,
-        tile_pos: IVec2,
-        chunk_pos: IVec2,
-    ) {
-        let obj_data = game
-            .chunk_manager
-            .chunk_generation_data
-            .get(&TileMapPositionData {
-                chunk_pos,
-                tile_pos: TilePos {
-                    x: tile_pos.x as u32,
-                    y: tile_pos.y as u32,
-                },
-            })
-            .unwrap();
+    // pub fn attempt_to_break_item(
+    //     self,
+    //     commands: &mut Commands,
+    //     game: &mut GameParam,
+    //     tile_pos: IVec2,
+    //     chunk_pos: IVec2,
+    // ) {
+    //     let obj_data = game
+    //         .chunk_manager
+    //         .chunk_generation_data
+    //         .get(&TileMapPositionData {
+    //             chunk_pos,
+    //             tile_pos: TilePos {
+    //                 x: tile_pos.x as u32,
+    //                 y: tile_pos.y as u32,
+    //             },
+    //         })
+    //         .unwrap();
 
-        let main_hand_tool = &game.game.player_state.main_hand_slot;
-        let b_data = game.block_query.get_mut(obj_data.entity).unwrap();
+    //     let main_hand_tool = &game.game.player_state.main_hand_slot;
+    //     let b_data = game.block_query.get_mut(obj_data.entity).unwrap();
 
-        if let Some(data) = game.world_obj_data.properties.get(&self) {
-            //TODO: maybe move this logic to combat.rs? or move the hp loss from combat to here?
-            if let Some(breaks_with) = data.breaks_with {
-                if let Some(main_hand_tool) = main_hand_tool {
-                    if main_hand_tool.obj == breaks_with {
-                        let h = b_data.1;
-                        if h.0 <= 0 {
-                            Self::break_item(self, commands, game, tile_pos, chunk_pos)
-                        }
-                    }
-                }
-            } else {
-                Self::break_item(self, commands, game, tile_pos, chunk_pos)
-            }
-        }
-    }
-    pub fn break_item(
-        self,
-        commands: &mut Commands,
-        game: &mut GameParam,
-        tile_pos: IVec2,
-        chunk_pos: IVec2,
-    ) {
-        let obj_data = game
-            .chunk_manager
-            .chunk_generation_data
-            .get(&TileMapPositionData {
-                chunk_pos,
-                tile_pos: TilePos {
-                    x: tile_pos.x as u32,
-                    y: tile_pos.y as u32,
-                },
-            })
-            .unwrap();
-
-        if let Some(breaks_into_option) = game.world_obj_data.properties.get(&self) {
-            commands.entity(obj_data.entity).despawn();
-            if let Some(breaks_into) = breaks_into_option.breaks_into {
-                let mut rng = rand::thread_rng();
-                breaks_into.spawn_item_drop(
-                    commands,
-                    game,
-                    tile_pos,
-                    chunk_pos,
-                    rng.gen_range(1..4),
-                    None,
-                );
-            }
-            game.chunk_manager
-                .chunk_generation_data
-                .remove(&TileMapPositionData {
-                    chunk_pos,
-                    tile_pos: TilePos {
-                        x: tile_pos.x as u32,
-                        y: tile_pos.y as u32,
-                    },
-                });
-            let old_points = game
-                .game_data
-                .data
-                .get(&(chunk_pos.x, chunk_pos.y))
-                .unwrap();
-            let updated_old_points = old_points
-                .0
-                .clone()
-                .iter()
-                .filter(|p| **p != (tile_pos.x as f32, tile_pos.y as f32, self))
-                .copied()
-                .collect::<Vec<(f32, f32, Self)>>();
-
-            game.game_data.data.insert(
-                (chunk_pos.x, chunk_pos.y),
-                ChunkObjectData(updated_old_points.to_vec()),
-            );
-        }
-    }
+    //     if let Some(data) = game.world_obj_data.properties.get(&self) {
+    //         //TODO: maybe move this logic to combat.rs? or move the hp loss from combat to here?
+    //         if let Some(breaks_with) = data.breaks_with {
+    //             if let Some(main_hand_tool) = main_hand_tool {
+    //                 if main_hand_tool.obj == breaks_with {
+    //                     let h = dbg!(b_data.1);
+    //                     if h.0 <= 0 {
+    //                         Self::break_item(self, commands, game, tile_pos, chunk_pos)
+    //                     }
+    //                 }
+    //             }
+    //         } else {
+    //             Self::break_item(self, commands, game, tile_pos, chunk_pos)
+    //         }
+    //     }
+    // }
 }
 
 impl Default for WorldObject {
@@ -724,12 +665,79 @@ impl Plugin for ItemsPlugin {
             .add_system_set(
                 SystemSet::on_update(GameState::Main)
                     .with_system(Self::update_graphics)
+                    .with_system(Self::break_item)
                     .with_system(Self::update_help_hotbar_item),
             );
     }
 }
 
 impl ItemsPlugin {
+    fn break_item(
+        mut commands: Commands,
+        mut game: GameParam,
+        mut obj_break_events: EventReader<ObjBreakEvent>,
+    ) {
+        // let obj_data = game
+        //     .chunk_manager
+        //     .chunk_generation_data
+        //     .get(&TileMapPositionData {
+        //         chunk_pos,
+        //         tile_pos: TilePos {
+        //             x: tile_pos.x as u32,
+        //             y: tile_pos.y as u32,
+        //         },
+        //     })
+        //     .unwrap();
+        for broken in obj_break_events.iter() {
+            if let Some(breaks_into_option) = game.world_obj_data.properties.get(&broken.obj) {
+                // commands.entity(obj_data.entity).despawn();
+                if let Some(breaks_into) = breaks_into_option.breaks_into {
+                    let mut rng = rand::thread_rng();
+                    println!("SPAWNING DROP");
+                    breaks_into.spawn_item_drop(
+                        &mut commands,
+                        &mut game,
+                        broken.tile_pos,
+                        broken.chunk_pos,
+                        rng.gen_range(1..4),
+                        None,
+                    );
+                }
+                game.chunk_manager
+                    .chunk_generation_data
+                    .remove(&TileMapPositionData {
+                        chunk_pos: broken.chunk_pos,
+                        tile_pos: TilePos {
+                            x: broken.tile_pos.x as u32,
+                            y: broken.tile_pos.y as u32,
+                        },
+                    });
+                let old_points = game
+                    .game_data
+                    .data
+                    .get(&(broken.chunk_pos.x, broken.chunk_pos.y))
+                    .unwrap();
+                let updated_old_points = old_points
+                    .0
+                    .clone()
+                    .iter()
+                    .filter(|p| {
+                        **p != (
+                            broken.tile_pos.x as f32,
+                            broken.tile_pos.y as f32,
+                            broken.obj,
+                        )
+                    })
+                    .copied()
+                    .collect::<Vec<(f32, f32, WorldObject)>>();
+
+                game.game_data.data.insert(
+                    (broken.chunk_pos.x, broken.chunk_pos.y),
+                    ChunkObjectData(updated_old_points.to_vec()),
+                );
+            }
+        }
+    }
     /// Keeps the graphics up to date for things that are harvested or grown
     fn update_graphics(
         mut to_update_query: Query<(&mut TextureAtlasSprite, &WorldObject), Changed<WorldObject>>,
